LayoutTests/imported/w3c/web-platform-tests/css/css-scroll-snap/snap-after-relayout/changing-scroll-snap-align-nested-expected.txt


FAIL Changing a large target's snap alignment shouldn't make the scroller resnap if the scroller is already in a valid snap position. assert_equals: expected 100 but got 200
PASS Changing the current (non-covering) target's snap alignment should make the scroller snap according to the new alignment.
